Revert "universal7880: Correct extra Samsung GPS field"
Confusion between int and long int.

This reverts commit 81aaae0fb545c6d6bce3c6f829a451f3ca29bf45.

+++ include/hardware/gps.h
@@ -571,7 +571,7 @@ typedef struct {
     /** Azimuth of SV in degrees. */
     float   azimuth;
     /** SAMSUNG */
-    int     samsung;
+    uint64_t samsung;
 } GpsSvInfo;
 
 typedef struct {
